requirements.txt
SQLAlchemy>=2.0.0,<3.0.0
bcrypt>=4.1.0,<5.0.0
Flask>=3.0.0,<4.0.0
Flask-Login>=0.6.3,<1.0.0
Flask-JWT-Extended>=4.6.0,<5.0.0
python-dotenv>=1.0.1,<2.0.0
gunicorn>=22.0.0,<23.0.0
psycopg2-binary>=2.9.9,<3.0.0
